refactor: Remove equals operators ('=') from braced initializer lists

+++ src/renderer.h
@@ -34,7 +34,7 @@ public:
     // Screen update and buffer swap is responsibility of window
 
 private:
-    enum class ShaderInstance : uint8_t
+    enum class ShaderInstance : size_t
     {
         ModelShader,
         SkyboxShader,
